:truck: api: move open api doc to shared
