+++ api/src/certification/results/infrastructure/open-api-doc/livret-scolaire/area-doc.js
