+++ api/src/certification/results/infrastructure/open-api-doc/livret-scolaire/certifications-results-doc.js
