+++ api/src/certification/results/infrastructure/open-api-doc/livret-scolaire/competence-doc.js
