+++ api/src/shared/infrastructure/open-api-doc/response-object-error-doc.js
